+++ atualizacao.h
@@ -9,7 +9,7 @@ Atualizacoes* initAtualizacoes();
 
 void liberaAtualizacoes(Atualizacoes* atualizacoes);
 
-void insereAtualizacao(Atualizacoes* atualizacoes, int origem, int destino, int instante, float velocidade);
+void insereAtualizacao(Atualizacoes* atualizacoes, int origem, int destino, int instante, double velocidade);
 
 int getOrigemAtualizacao(Atualizacoes* atualizacoes, int i);
 
@@ -17,7 +17,7 @@ int getDestinoAtualizacao(Atualizacoes* atualizacoes, int i);
 
 int getInstanteAtualizacao(Atualizacoes* atualizacoes, int i);
 
-float getVelocidadeAtualizacao(Atualizacoes* atualizacoes, int i);
+double getVelocidadeAtualizacao(Atualizacoes* atualizacoes, int i);
 
 Atualizacao** getVetorAtualizacao(Atualizacoes* atualizacoes);
